feat(update-user): update user profile api endpoint
- Code Refactoring
- Used multer and cloudinary for handling images
- modify database fields

[Finishes #159728482]

+++ server/controllers/categories.js
@@ -1,6 +1,19 @@
 import client from '../models/database/dbconnect';
 
+/**
+ * Class implementation for /api/v1/user routes
+ * @class CategoriesController
+ */
+
 class CategoriesController {
+  /**
+   * @description - Creating new Category
+   *
+   * @param { object }  req
+   * @param { object }  res
+   *
+   * @returns { object } object
+   */
   async addCategory(req, res) {
     try {
       const {
@@ -8,24 +21,30 @@ class CategoriesController {
         colorId,
       } = req.body;
 
+      const userId = req.userData.userID;
+
       const checkTitle = await client.query(
-        `SELECT * FROM categories WHERE title=($1)`, [
+        `SELECT * FROM categories WHERE title=($1) AND user_id=($2)`, [
           title,
+          userId,
         ],
       );
 
       if (!checkTitle.rows.length > 0) {
-        const sql = `INSERT INTO categories(title,color_id) VALUES($1, $2)`;
+        const sql = `INSERT INTO categories
+          (title,color_id, user_id) VALUES($1, $2, $3)`;
         const values = [
           title.trim(),
           colorId,
+          userId,
         ];
 
         await client.query(sql, values);
 
         const fetchCategory = await client.query(
-          `SELECT * FROM categories WHERE title=($1)`, [
+          `SELECT * FROM categories WHERE title=($1) AND user_id=($2)`, [
             title,
+            userId,
           ],
         );
         const newCategory = fetchCategory.rows;
@@ -41,11 +60,19 @@ class CategoriesController {
     } catch (error) {
       return res.status(500).json({
         message: 'Error processing request',
-        error,
+        error: error.toString(),
       });
     }
   }
 
+  /**
+   * @description - Deleting category
+   *
+   * @param { object }  req
+   * @param { object }  res
+   *
+   * @returns { object } object
+   */
   async deleteCategory(req, res) {
     try {
       const {
@@ -61,7 +88,7 @@ class CategoriesController {
       }
 
       const fetchCategory = await client.query(
-        `DELETE FROM categories WHERE category_id=($1)`, [
+        `DELETE FROM categories WHERE category_id=($1), user_id=($2)`, [
           categoryid,
         ],
       );
@@ -83,10 +110,19 @@ class CategoriesController {
     } catch (error) {
       return res.status(500).json({
         message: `Error Processing Request`,
-        error,
+        error: error.toString(),
       });
     }
   }
+
+  /**
+   * @description - Fetching all categories
+   *
+   * @param { object }  req
+   * @param { object }  res
+   *
+   * @returns { object } object
+   */
 
   async getAllCategories(req, res) {
     try {
@@ -103,10 +139,19 @@ class CategoriesController {
     } catch (error) {
       return res.status(500).json({
         message: `Error Processing Request`,
-        error,
+        error: error.toString(),
       });
     }
   }
+
+  /**
+   * @description - Fetch a single category
+   *
+   * @param { object }  req
+   * @param { object }  res
+   *
+   * @returns { object } object
+   */
 
   async getSingleCategory(req, res) {
     try {
@@ -139,7 +184,7 @@ class CategoriesController {
     } catch (error) {
       return res.status(500).json({
         message: `Error processing request.`,
-        error,
+        error: error.toString(),
       });
     }
   }
